# HtmlStream: include element range in parse_with_range