--- a/htmlStream_ragel.ml.rl
+++ b/htmlStream_ragel.ml.rl
@@ -13,7 +13,9 @@ type elem =
 | Text of Raw.t
 | Close of string
 
-type ctx = { mutable lnum : int }
+type range = (int * int)
+
+type ctx = { mutable lnum : int; }
 
 let get_lnum ctx = ctx.lnum
 
@@ -24,20 +26,21 @@ let init () = { lnum = 1 }
 
  action mark { mark := !p }
  action mark_end { mark_end := !p }
- action tag { tag := String.lowercase_ascii @@ sub (); attrs := []; }
- action close_tag { call @@ Close (String.lowercase_ascii @@ sub ()) }
- action directive { directive := String.lowercase_ascii @@ sub (); attrs := []; }
- action text { call @@ Text (Raw.inject @@ sub ()) }
- action key { key := String.lowercase_ascii @@ sub () }
- action store_attr { attrs := (!key, Raw.inject (if !mark < 0 then "" else sub())) :: !attrs }
+ action tag_start { tag_start := !p }
+ action tag { let s, _r = sub () in tag := String.lowercase_ascii s; attrs := []; }
+ action close_tag { let s, _r = sub () in call @@ (Close (String.lowercase_ascii s), tag_range ()) }
+ action directive { let s, _r = sub () in directive := (String.lowercase_ascii s); attrs := []; }
+ action text { let s, r = sub () in call @@ (Text (Raw.inject s), r) }
+ action key { let s, _r = sub () in key := String.lowercase_ascii s; }
+ action store_attr { attrs := (!key, Raw.inject (if !mark < 0 then "" else let s, _r = sub() in s)) :: !attrs }
  action tag_done {
     match !tag with
     | "script" -> fhold; fgoto in_script;
     | "style" -> fhold; fgoto in_style;
     | "" -> ()
-    | _ -> call @@ Tag (!tag, List.rev !attrs)
+    | _ -> call @@ (Tag (!tag, List.rev !attrs), tag_range ())
  }
- action tag_done_2 { call @@ Tag (!tag, List.rev !attrs); if !tag <> "a" then call (Close !tag) }
+ action tag_done_2 { let r = tag_range () in call @@ (Tag (!tag, List.rev !attrs), r); if !tag <> "a" then call ((Close !tag), r) }
  action directive_done { (* printfn "directive %s" !directive; *) }
 
  action garbage_tag { (*printfn "GARBAGE %S" (current ()); *) fhold; fgoto garbage_tag;}
@@ -47,29 +50,34 @@ let init () = { lnum = 1 }
  wsp = 0..32;
  ident = alnum | '-' | [_:] ;
 
- in_script := (count_newlines | any* >mark %mark_end :>> ('<' wsp* '/' wsp* 'script'i wsp* '>' >{call @@ Script (List.rev !attrs, sub ())} @{fgoto main;}));
- in_style := (count_newlines | any* >mark %mark_end :>> ('<' wsp* '/' wsp* 'style'i wsp* '>' >{call @@ Style (List.rev !attrs, sub ())} @{fgoto main;}));
+ in_script := (count_newlines | any* >mark %mark_end :>> ('<' wsp* '/' wsp* 'script'i wsp* '>' >{
+  let s, _r = sub () in call @@ (Script (List.rev !attrs, s), tag_range ())
+ } @{fgoto main;}));
+ in_style := (count_newlines | any* >mark %mark_end :>> ('<' wsp* '/' wsp* 'style'i wsp* '>' >{
+  let s, _r = sub () in call @@ (Style (List.rev !attrs, s), tag_range ())
+ } @{fgoto main;}));
 
  garbage_tag := (count_newlines | ^'>'* '>' @tag_done @{ fgoto main; });
 
  literal = ( "'" ^"'"* >mark %mark_end "'" | '"' ^'"'* >mark %mark_end '"' | ^(wsp|'"'|"'"|'>')+ >mark %mark_end);
  tag_attrs = (wsp+ | ident+ >mark %key wsp* ('=' wsp* literal)? %store_attr )**;
- close_tag = '/' wsp* ident* >mark %close_tag <: ^'>'* '>';
+ close_tag = '/' wsp* ident* >mark %mark_end <: ^'>'* '>' %close_tag;
  open_tag = ident+ >mark %tag <: wsp* tag_attrs ('/' wsp* '>' %tag_done_2 | '>' %tag_done);
  directive = ('!'|'?') (alnum ident+) >mark %directive <: wsp* tag_attrs '?'? '>' %directive_done;
  comment = "!--" any* :>> "-->";
  # reset tag so that garbage_tag will not generate duplicate tag with tag_done
- tag = '<' wsp* <: (close_tag | open_tag | directive | comment) @lerr(garbage_tag) >{ tag := "" };
+ tag = '<' >tag_start wsp* <: (close_tag | open_tag | directive | comment) @lerr(garbage_tag) >{ tag := "" };
  main := (((tag | ^'<' >mark ^'<'* %text ) )** | count_newlines);
 
  write data;
 }%%
 
 (** scan [data] for html tags and invoke [call] for every element  *)
-let parse ?(ctx=init ()) call data =
+let parse_with_range ?(ctx=init ()) call data =
   let cs = ref 0 in
   let mark = ref (-1) in
   let mark_end = ref (-1) in
+  let tag_start = ref (-1) in 
   let tag = ref "" and key = ref "" and attrs = ref [] and directive = ref "" in
 (*  let substr data ofs len = try String.sub data ofs len with exn -> Prelude.printfn "%S %d %d %d" data (String.length data) ofs len; raise exn in *)
   let substr = String.sub in
@@ -81,9 +89,16 @@ let parse ?(ctx=init ()) call data =
     assert (!mark >= 0);
     if !mark_end < 0 then mark_end := !p;
     let s = if !mark_end <= !mark then "" else substr data !mark (!mark_end - !mark) in
+    let range = (!mark, !mark_end) in
     mark := -1;
     mark_end := -1;
-    s
+    (s, range)
+  in
+  let tag_range () = 
+    assert (!tag_start >= 0);
+    let range = (!tag_start, !p) in
+    tag_start := -1;
+    range
   in
   %%write exec;
 (* FIXME ? *)