This PR adds a `parse_with_range` function under `HtmlStream` which includes the element's position (start+end offset) when calling the callback. 

Also included some tests that checks the output offsets are correct by comparing substring on the HTML input vs the parsed element.