@tysg 2023-08-01
## Benchmarks 
### Setup
A ~400k `yahoo.html` file:
```sh
$ ~/a/devkit (master)> ll yahoo.html
Permissions Size User Date Modified Name
.rw-r--r--  418k me    1 Aug  3:01  yahoo.html
```
Code:
```ocaml
open Devkit
let () = 
  let n = int_of_string @@ Sys.argv.(1) in
  let s = In_channel.with_open_text "yahoo.html" In_channel.input_all in
  Action.run_bench n
  [
    ( "parse", fun () -> HtmlStream.parse ignore s;)
  ]
```
### Baseline (current master)
```sh
$ ~/a/devkit (master)> git show | head -n 1
commit 3239108737de6e4169de04c666740ee1ab7a4755 (HEAD -> master, origin/master, origin/HEAD)
$ ~/a/devkit (master)> dune exec ./htmlstream_bench.exe -- 1
run_bench 1 cases (count 1)            
parse : allocated      635KB, heap         0B, collection 0 0 0, elapsed 0.0056 seconds, 177.43/sec : ok
$ ~/a/devkit (master)> dune exec ./htmlstream_bench.exe -- 10
run_bench 1 cases (count 10)           
parse : allocated      6.2MB, heap      2.8MB, collection 0 0 1, elapsed 0.0571 seconds, 175.24/sec : ok
```
### This PR
```sh
$ ~/a/devkit (htmlstream-parse-with-range)> git show | head -n 1
commit ef8d38b76047db4ddf3cf0e56a62d6149928702f (HEAD -> htmlstream-parse-with-range)
$ ~/a/devkit (htmlstream-parse-with-range)> dune exec ./htmlstream_bench.exe -- 1
run_bench 1 cases (count 1)            
parse : allocated      908KB, heap         0B, collection 0 0 0, elapsed 0.0057 seconds, 176.49/sec : ok
$ ~/a/devkit (htmlstream-parse-with-range)> dune exec ./htmlstream_bench.exe -- 10
run_bench 1 cases (count 10)           
parse : allocated      8.9MB, heap      2.8MB, collection 0 0 2, elapsed 0.0573 seconds, 174.41/sec : ok
```
We see a 50% increase in allocated memory (largely because of the additional tuple) but very little change in processing time. 

I also attempted to reduce the number of tuples allocated by changing `call` from `(elem, range) -> unit` to `elem -> range -> unit`, and only see minor improvement in allocation (see details).

<details>

```sh
$ ~/a/devkit (htmlstream-parse-with-range)> git diff htmlStream_ragel.ml.rl | head -n 10
diff --git a/htmlStream_ragel.ml.rl b/htmlStream_ragel.ml.rl
index 0968488..6e7a98b 100644
--- a/htmlStream_ragel.ml.rl
+++ b/htmlStream_ragel.ml.rl
@@ -28,9 +28,9 @@ let init () = { lnum = 1 }
  action mark_end { mark_end := !p }
  action tag_start { tag_start := !p }
  action tag { let s, _r = sub () in tag := String.lowercase_ascii s; attrs := []; }
- action close_tag { let s, _r = sub () in call @@ (Close (String.lowercase_ascii s), tag_range ()) }
+ action close_tag { let s, _r = sub () in call (Close (String.lowercase_ascii s)) (tag_range ()) }
$ ~/a/devkit (htmlstream-parse-with-range)> dune exec ./htmlstream_bench.exe -- 10
run_bench 1 cases (count 10)           
parse : allocated      8.5MB, heap      2.8MB, collection 0 0 2, elapsed 0.0582 seconds, 171.96/sec : ok
```

</details>